% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_helpers.R
\name{create_trial_numbers}
\alias{create_trial_numbers}
\title{Create trial numbers following L2T WordList conventions}
\usage{
create_trial_numbers(trial_types)
}
\arguments{
\item{trial_types}{a character vector of the words "Familiarization" and
"Test" (which indicated whether a trial is a warm-up or test trial)}
}
\value{
a vector of L2T WordList trial numbers
}
\description{
Create trial numbers following L2T WordList conventions
}
\details{
"Familiarization" is abbreviated to "Fam". Then trial types are
numbered separately. For example, `c("Familiarization", "Familiarization",
"Test", "Test", "Test")` would output `c("Fam1", "Fam2", "Test1", "Test2",
"Test3")`
}

