% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwr_helpers.R, R/rwr_helpers.R
\name{create_nwr_wordlist_file}
\alias{create_nwr_wordlist_file}
\alias{create_rwr_wordlist_file}
\title{Convenience function to create a WordList given a ResearchID, study and
folder}
\usage{
create_nwr_wordlist_file(short_id, study_name, dir_task,
  dir_eprime = "Recordings", dir_wordlist = "WordLists", read_only = TRUE,
  update = FALSE)

create_rwr_wordlist_file(short_id, study_name, dir_task,
  dir_eprime = "Recordings", dir_wordlist = "WordLists", read_only = TRUE,
  update = FALSE)
}
\arguments{
\item{short_id}{four-character short participant ID (e.g., "001L"). Use `""`
for a wildcard search that will match all 4-character participants IDs.}

\item{study_name}{name of an L2T study (e.g., "TimePoint1")}

\item{dir_task}{the fully specified filepath to the folder for a Task. The
value for `study_name` should be a subfolder in this location.}

\item{dir_eprime}{folder that contains Eprime output files. Defaults to
`"Recordings"`.}

\item{dir_wordlist}{folder where the WordList should be saved. Defaults to
`"WordLists"`.}

\item{read_only}{whether to protect the saved WordList. Defaults to `TRUE`.}

\item{update}{whether to update the saved WordList. Defaults to `FALSE`.}
}
\value{
a list of the file-paths of the generated WordList files and the
  contents of the WordLists files (as data-frames).
}
\description{
Convenience function to create a WordList given a ResearchID, study and
folder
}

