% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{extract_l2t_id}
\alias{extract_l2t_id}
\title{Locate and extract an L2T id from a string or vector of strings}
\usage{
extract_l2t_id(xs)
}
\description{
Locate and extract an L2T id from a string or vector of strings
}

