% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwr_helpers.R
\name{get_nwr_trial_info}
\alias{get_nwr_trial_info}
\title{Extract trial-level data from a NWR Eprime text file}
\usage{
get_nwr_trial_info(eprime_path)
}
\arguments{
\item{eprime_path}{path to an Eprime text file}
}
\value{
a data-frame with information about each trial
}
\description{
Extract trial-level data from a NWR Eprime text file
}

