% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwr_helpers.R
\name{get_rwr_trial_info}
\alias{get_rwr_trial_info}
\title{Extract trial-level data from a RWR Eprime text file}
\usage{
get_rwr_trial_info(eprime_path)
}
\arguments{
\item{eprime_path}{path to an Eprime text file}
}
\value{
a data-frame with information about each trial
}
\description{
Extract trial-level data from a RWR Eprime text file
}

