% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwr_helpers.R
\name{get_wordlist_info}
\alias{get_wordlist_info}
\title{Create a WordList for a RWR expriment}
\usage{
get_wordlist_info(df_trials)
}
\arguments{
\item{df_trials}{a data-frame with trial-level information about a RWR
repetition experiment}
}
\value{
a data-frame with the "WordList" for those trials
}
\description{
Create a WordList for a RWR expriment
}

