% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{l2t_wordlists}
\alias{l2t_wordlists}
\title{Stimulus descriptions for L2T repetition experiments}
\format{An object of class \code{list} of length 2.}
\usage{
l2t_wordlists
}
\description{
Stimulus descriptions for L2T repetition experiments
}
\keyword{datasets}

