% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwr_helpers.R
\name{lookup_nwr_wordlist}
\alias{lookup_nwr_wordlist}
\title{Create a WordList for a NWR expriment}
\usage{
lookup_nwr_wordlist(df_trials)
}
\arguments{
\item{df_trials}{a data-frame with trial-level information about a NWR
repetition experiment}
}
\value{
a data-frame with the "WordList" for those trials
}
\description{
Create a WordList for a NWR expriment
}

