% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_protected_tsv}
\alias{write_protected_tsv}
\title{Write a read-only tab-separated-value file}
\usage{
write_protected_tsv(x, path)
}
\arguments{
\item{x}{a dataframe}

\item{path}{path to where the file should be saved}
}
\value{
invisibly returns the input `x`
}
\description{
Write a read-only tab-separated-value file
}

